Move tsn qos-related structs to alinx_arch
